* Fix ffmpeg timestamp generation * Misc other fixes
git-svn-id: svn://svn.code.sf.net/p/gmerlin/code/trunk@2650 214d8ac5-8513-4f59-80f8-2f39b674a24a

+++ gmerlin_avdecoder/lib/ptscache.c
@@ -27,7 +27,7 @@ static int get_min_index(bgav_pts_cache_t * c)
   {
   int i;
   int ret = -1;
-  int min;
+  int64_t min;
   for(i = 0; i < PTS_CACHE_SIZE; i++)
     {
     if(c->entries[i].used)
